+++ src/commands/PlayoffCommands.ts
@@ -7,30 +7,12 @@ import { PlayoffRoundFormatter } from "../utils/EmbedFormatters";
 export const GetPlayoffStandings: Command = {
 	name: 'playoffs',
 	description: 'Current playoff standings',
-	help: 'playoffs',
-	async execute(message: Message, args?: string[]) {
-        const rounds = await API.Playoffs.GetPlayoffStandings();
-        if(!rounds?.[0]){
-            message.channel.send('Playoff results not available.')
-            return;
-        }
-	
-	    message.channel.send({embeds: [
-            new EmbedBuilder({
-                title: 'Playoff Standings',
-                fields: rounds.map(round => {
-                    return PlayoffRoundFormatter(round);
-                })
-            })
-        ]});
-
-	},
-	slashCommandDescription: () => {
-		return new SlashCommandBuilder()
-		.setName('playoffs')
-		.setDescription('Current playoff standings');
-	},
-	executeSlashCommand: async (interaction) => {
+    slashCommandDescription: () => {
+        return new SlashCommandBuilder()
+        .setName('playoffs')
+        .setDescription('Current playoff standings');
+    },
+    executeSlashCommand: async (interaction) => {
         interaction.deferReply();
         const rounds = await API.Playoffs.GetPlayoffStandings();
         if(!rounds?.[0]){
@@ -47,5 +29,21 @@ export const GetPlayoffStandings: Command = {
         ]});
         
         
-	}
+    }
+}
+const oldCommand = async (message: Message, args?: string[]) => {
+    const rounds = await API.Playoffs.GetPlayoffStandings();
+    if(!rounds?.[0]){
+        message.channel.send('Playoff results not available.')
+        return;
+    }
+
+    message.channel.send({embeds: [
+        new EmbedBuilder({
+            title: 'Playoff Standings',
+            fields: rounds.map(round => {
+                return PlayoffRoundFormatter(round);
+            })
+        })
+    ]});
 }
